Attempt drain using liquid or lightning payment

+++ lib/core/src/model.rs
@@ -100,9 +100,9 @@ impl Config {
             .unwrap_or(DEFAULT_ZERO_CONF_MAX_SAT)
     }
 
-    pub(crate) fn lowball_fee_rate_msat_per_vbyte(&self) -> Option<f64> {
+    pub(crate) fn lowball_fee_rate_msat_per_vbyte(&self) -> Option<f32> {
         match self.network {
-            LiquidNetwork::Mainnet => Some(LOWBALL_FEE_RATE_SAT_PER_VBYTE * 1000.0),
+            LiquidNetwork::Mainnet => Some((LOWBALL_FEE_RATE_SAT_PER_VBYTE * 1000.0) as f32),
             LiquidNetwork::Testnet => None,
         }
     }
